Moved more component based state logic into redux based.
